refactor: 자외선 레벨을 리턴하도록 리팩토링

+++ src/main/java/capsrock/ultraviolet/dto/service/Dashboard.java
@@ -4,6 +4,6 @@ import capsrock.location.geocoding.dto.service.AddressDTO;
 
 public record Dashboard(
         AddressDTO address,
-        Double ultravioletLevel
+        Integer ultravioletLevel
 ) {
 }
